refactor: move version and prog back to __init__.py to keep similar project style
